--- a/cmd/sim/go.mod
+++ b/cmd/sim/go.mod
@@ -4,9 +4,14 @@ go 1.20
 
 replace github.com/eatonphil/goraft => ../../
 
-require github.com/eatonphil/goraft v0.0.0-00010101000000-000000000000
+require (
+	github.com/eatonphil/goraft v0.0.0-00010101000000-000000000000
+	github.com/pkg/profile v1.7.0
+)
 
 require (
+	github.com/felixge/fgprof v0.9.3 // indirect
+	github.com/google/pprof v0.0.0-20211214055906-6f57359322fd // indirect
 	github.com/petermattis/goid v0.0.0-20180202154549-b0b1615b78e5 // indirect
 	github.com/sasha-s/go-deadlock v0.3.1 // indirect
 )