--- a/raft.go
+++ b/raft.go
@@ -12,8 +12,8 @@ import (
 	"net/rpc"
 	"os"
 	"path"
-	"sync"
-	//sync "github.com/sasha-s/go-deadlock"
+	//"sync"
+	sync "github.com/sasha-s/go-deadlock"
 	"time"
 )
 
@@ -233,7 +233,7 @@ func NewServer(
 	metadataDir string,
 	clusterIndex int,
 ) *Server {
-	//sync.Opts.DeadlockTimeout = 50 * time.Millisecond
+	//sync.Opts.DeadlockTimeout = 2000 * time.Millisecond
 	// Explicitly make a copy of the cluster because we'll be
 	// modifying it in this server.
 	var cluster []ClusterMember
@@ -260,7 +260,7 @@ func NewServer(
 
 const PAGE_SIZE = 4096
 const ENTRY_HEADER = 16
-const ENTRY_SIZE = PAGE_SIZE
+const ENTRY_SIZE = 256
 
 // Weird thing to note is that writing to a deleted disk is not an
 // error on Linux. So if these files are deleted, you won't know about