--- a/raft.go
+++ b/raft.go
@@ -12,8 +12,8 @@ import (
 	"net/rpc"
 	"os"
 	"path"
-	//"sync"
-	sync "github.com/sasha-s/go-deadlock"
+	"sync"
+	//sync "github.com/sasha-s/go-deadlock"
 	"time"
 )