* Add support for batch Apply
* Debounce appendEntries when called from Apply
* Add support for testing with multiple clients from cmd/sim
* Batch append log writes with bufio.Writer

Before:

```
Waiting for a leader.
1000 entries inserted in 19.209651083s.
2000 entries inserted in 19.715948417s.
Total time: 51.371829178s. Average insert/second: 17.123943ms. Throughput: 58.823529 entries/s.
eUWYXFOYDXZJoTCk = BnOtVMyDZGLdiGFs, expected: BnOtVMyDZGLdiGFs
```

After:

```
Waiting for a leader.
Clients: 1. Entries: 3000. Batch: 8000.
Client: 0. 3000 entries (3000 of 3000: 100%) inserted. Latency: 3.236311292s. Throughput: 1000.000000 entries/s.
Total time: 3.236311292s. Average insert/second: 1.07877ms. Throughput: 1000.000000 entries/s.
Waiting for commits to be applied.
QJHcLMQONqgoDKPv = hlpLYHPsqwBtSehj, expected: hlpLYHPsqwBtSehj
```